modify, add: MainViewController / OptionViewController - modified : 메인뷰 새로운 레이아웃으로 변경 - added : 옵션뷰 추가
