+++ NetHawk/NetHawk/Controllers/MainViewController.swift
@@ -1,193 +1,105 @@
 //
-//  ViewController.swift
-//  PackeTracker
+//  MainViewController2.swift
+//  NetHawk
 //
-//  Created by mobicom on 6/4/24.
+//  Created by mobicom on 9/10/24.
 //
 
 import UIKit
-import CoreLocation
-import SystemConfiguration
+import FSPagerView
 
-class MainViewController: UIViewController, CLLocationManagerDelegate {
-    
-    // MARK: - UI Outlets
-    @IBOutlet weak var titleBtn: UIButton!
-    @IBOutlet weak var alertBtn: UIButton!
-    
-    @IBOutlet weak var logBtn: UIButton!
-    @IBOutlet weak var statsBtn: UIButton!
-    @IBOutlet weak var ipBtn: UIButton!
-    
-    @IBOutlet weak var wifiStatusTitle: UILabel!
-    @IBOutlet weak var wifiStatusLabel: UILabel!
-    
-    @IBOutlet weak var logLabel: UILabel!
-    @IBOutlet weak var statsLabel: UILabel!
-    @IBOutlet weak var ipPortLabel: UILabel!
-    
-    @IBOutlet weak var logo: UIImageView!
-    // MARK: - Properties
-    let locationManager = CLLocationManager()
-    let wifiService = WiFiService()
-    
-    // MARK: - Life Cycles
+class MainViewController: UIViewController, FSPagerViewDataSource, FSPagerViewDelegate {
+
+    @IBOutlet weak var pagerView: FSPagerView! {
+        didSet {
+            self.pagerView.register(FSPagerViewCell.self, forCellWithReuseIdentifier: "cell")
+            self.pagerView.itemSize = CGSize(width:220, height: 150)
+            self.pagerView.interitemSpacing = 16
+            self.pagerView.isInfinite = true // 무한 스크롤 가능
+            self.pagerView.transformer = FSPagerViewTransformer(type: .ferrisWheel)
+        }
+    }
+
+    let images = ["list.bullet.clipboard", "chart.bar.xaxis", "shield.lefthalf.filled.slash", "gearshape"]
+    let titles = ["Log", "Statistics", "Black/White List", "Option"]
+
     override func viewDidLoad() {
         super.viewDidLoad()
-        print("----------------")
-        print("viewDidLoad : MainView")
-        setupTitleButton()
-        setupWifiStatusLabel()
-        
-        applyButtonEffect(logBtn)
-        applyButtonEffect(statsBtn)
-        applyButtonEffect(ipBtn)
-        
-        addShadowToButton(logBtn)
-        addShadowToButton(statsBtn)
-        addShadowToButton(ipBtn)
-        
-        locationManager.delegate = self
-        locationManager.requestWhenInUseAuthorization()
-        
-        titleBtn.alpha = 0.0
-        alertBtn.alpha = 0.0
-        logBtn.alpha = 0.0
-        statsBtn.alpha = 0.0
-        ipBtn.alpha = 0.0
-        wifiStatusLabel.alpha = 0.0
-        wifiStatusTitle.alpha = 0.0
-        logLabel.alpha = 0.0
-        statsLabel.alpha = 0.0
-        ipPortLabel.alpha = 0.0
-        logo.alpha = 0.0
+        self.pagerView.dataSource = self
+        self.pagerView.delegate = self
     }
-    
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        // 0.1초 지연 후에 애니메이션 시작
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
-            self.applyAnimations()
+
+    // FSPagerViewDataSource methods
+    func numberOfItems(in pagerView: FSPagerView) -> Int {
+        return images.count // 총 4개의 페이지
+    }
+
+    func pagerView(_ pagerView: FSPagerView, cellForItemAt index: Int) -> FSPagerViewCell {
+        let cell = pagerView.dequeueReusableCell(withReuseIdentifier: "cell", at: index)
+
+        // SF Symbol 이미지 설정
+        if let imageView = cell.imageView {
+            imageView.image = UIImage(systemName: images[index])
+            imageView.contentMode = .scaleAspectFit // 이미지의 비율을 유지하면서 셀에 맞춤
         }
-    }
-    
-    
-    // MARK: - Layout Config
-    private func setupTitleButton() {
-        titleBtn.setTitle(".NETHAWK", for: .normal)
-        titleBtn.layer.shadowColor = UIColor.gray.cgColor
-        titleBtn.layer.shadowOpacity = 0.4
-        titleBtn.layer.shadowOffset = CGSize(width: 4, height: 3)
-        titleBtn.layer.shadowRadius = 4
-        titleBtn.translatesAutoresizingMaskIntoConstraints = false
-    }
-    
-    private func setupWifiStatusLabel() {
-        wifiStatusLabel.textAlignment = .center
-        wifiStatusLabel.numberOfLines = 0
-        wifiStatusLabel.font = UIFont(name: "IntelOneMono-Light", size: 20)
-    }
-    
-    private func addShadowToButton(_ button: UIButton) {
-        button.layer.shadowColor = UIColor.black.cgColor
-        button.layer.shadowOffset = CGSize(width: 0, height: 2)
-        button.layer.shadowOpacity = 0.5
-        button.layer.shadowRadius = 4
-        button.layer.masksToBounds = false
-    }
-    
-    func applyAnimations() {
-        UIView.animate(withDuration: 1, delay: 0.1, options: .curveEaseInOut, animations: {
-            self.logo.alpha = 0.15
-            
-        }, completion: nil)
-        
-        UIView.animate(withDuration: 1, delay: 0.5, options: .curveEaseInOut, animations: {
-            self.titleBtn.alpha = 1.0
-            self.alertBtn.alpha = 1.0
 
-        }, completion: nil)
-        
-        UIView.animate(withDuration: 1, delay: 0.7, options: .curveEaseInOut, animations: {
-            self.wifiStatusLabel.alpha = 1.0
-            self.wifiStatusTitle.alpha = 1.0
+        // 레이블 설정
+        if let textLabel = cell.textLabel {
+            textLabel.text = titles[index]
+            textLabel.textAlignment = .center
+            textLabel.textColor = .darkGray
+            textLabel.font = UIFont(name: "IntelOneMono-Medium", size: 16)
 
-        }, completion: nil)
+            if let superview = textLabel.superview {
+                superview.backgroundColor = UIColor.clear // 배경색을 투명하게 설정
+                superview.translatesAutoresizingMaskIntoConstraints = false
 
-        
-        UIView.animate(withDuration: 1, delay: 0.9, options: .curveEaseInOut, animations: {
-            self.logBtn.alpha = 1.0
-            self.logLabel.alpha = 1.0
-            self.statsBtn.alpha = 1.0
-            self.statsLabel.alpha = 1.0
-            self.ipBtn.alpha = 1.0
-            self.ipPortLabel.alpha = 1.0
-        }, completion: nil)
+                // 슈퍼뷰의 새로운 제약 조건 설정 (아래로 20 포인트 이동)
+                NSLayoutConstraint.activate([
+                    superview.topAnchor.constraint(equalTo: superview.superview!.topAnchor, constant: 140), // 상단에서 20포인트 떨어짐
+                    superview.leadingAnchor.constraint(equalTo: superview.superview!.leadingAnchor),
+                    superview.trailingAnchor.constraint(equalTo: superview.superview!.trailingAnchor),
+                    superview.heightAnchor.constraint(equalToConstant: textLabel.frame.size.height)
+                ])
+            }
+        }
+
+        return cell
     }
-    
-    // MARK: - Alert [위치동의]
-    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-        switch status {
-        case .authorizedWhenInUse, .authorizedAlways:
-            print("Location permission granted")
-            updateWiFiInfo()
+
+    func pagerView(_ pagerView: FSPagerView, didSelectItemAt index: Int) {
+
+        pagerView.deselectItem(at: index, animated: true) // 선택된 상태 해제
+        print("Selected page index: \(index)")
+
+        let storyboard = UIStoryboard(name: "Main", bundle: nil)
+        switch index {
+        case 0:
+            // 첫 번째 페이지로 이동
+            if let firstVC = storyboard.instantiateViewController(withIdentifier: "LogViewController") as? LogViewController {
+                firstVC.modalPresentationStyle = .fullScreen
+                present(firstVC, animated: true, completion: nil)
+            }
+        case 1:
+            // 두 번째 페이지로 이동
+            if let secondVC = storyboard.instantiateViewController(withIdentifier: "StatViewController") as? StatViewController {
+                secondVC.modalPresentationStyle = .fullScreen
+                present(secondVC, animated: true, completion: nil)
+            }
+        case 2:
+            // 세 번째 페이지로 이동
+            if let thirdVC = storyboard.instantiateViewController(withIdentifier: "AccessViewController") as? AccessViewController {
+                thirdVC.modalPresentationStyle = .fullScreen
+                present(thirdVC, animated: true, completion: nil)
+            }
+        case 3:
+            // 세 번째 페이지로 이동
+            if let fourVC = storyboard.instantiateViewController(withIdentifier: "OptionViewController") as? OptionViewController {
+                fourVC.modalPresentationStyle = .fullScreen
+                present(fourVC, animated: true, completion: nil)
+            }
         default:
-            print("Location permission not granted")
-            wifiStatusLabel.text = "Location permission not granted"
-        }
-    }
-    
-    func updateWiFiInfo() {
-        print("updateWiFiInfo: Checking WiFi information")
-        if let ipAddress = wifiService.getWiFiAddress() {
-            print("updateWiFiInfo: IP address found - \(ipAddress)")
-            wifiStatusLabel.text = "IP : \(ipAddress)"
-        } else {
-            print("updateWiFiInfo: No WiFi connection found")
-            wifiStatusLabel.text = "WiFi 연결 안됨"
-        }
-    }
-    
-    
-    @IBAction func logBtnTapped(_ sender: UIButton) {
-        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        if let logVC = storyboard.instantiateViewController(withIdentifier: "LogViewController") as? LogViewController {
-            logVC.modalPresentationStyle = .fullScreen
-            present(logVC, animated: true, completion: nil)
-        }
-    }
-
-    @IBAction func statBtntapped(_ sender: UIButton) {
-        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        if let statVC = storyboard.instantiateViewController(withIdentifier: "StatViewController") as? StatViewController {
-            statVC.modalPresentationStyle = .fullScreen
-            present(statVC, animated: true, completion: nil)
-        }
-    }
-    
-    @IBAction func ipBtnTapped(_ sender: UIButton) {
-        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        if let accessVC = storyboard.instantiateViewController(withIdentifier: "AccessViewController") as? AccessViewController {
-            accessVC.modalPresentationStyle = .fullScreen
-            present(accessVC, animated: true, completion: nil)
-        }
-    }
-    
-    
-    func applyButtonEffect(_ button: UIButton) {
-        button.addTarget(self, action: #selector(buttonTouchDown(_:)), for: .touchDown)
-        button.addTarget(self, action: #selector(buttonTouchUp(_:)), for: [.touchUpInside, .touchUpOutside, .touchCancel])
-    }
-    
-    @objc func buttonTouchDown(_ sender: UIButton) {
-        UIView.animate(withDuration: 0.1) {
-            sender.alpha = 0.5
-        }
-    }
-    
-    @objc func buttonTouchUp(_ sender: UIButton) {
-        UIView.animate(withDuration: 0.4) {
-            sender.alpha = 1.0
+            break
         }
     }
 }
