+++ NetHawk/NetHawk/Controllers/OptionViewController.swift
@@ -11,8 +11,11 @@ class OptionViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
+        print("viewDidLoad : OptionView")
     }
-    
+
+    @IBAction func dissmissBtnTapped(_ sender: UIButton) {
+        dismiss(animated: true, completion: nil)
+    }
 }
